add lab01

+++ lab02-3.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 #include <math.h>
 
-int factorial(int n) {
+double factorial(double n) {
 	if (n == 0) {
-		return 1;
+		return 1.0;
 	}
-	return (n * factorial(n - 1));
+	return (n * factorial(n - 1.0));
 }
 double myexp(int x) {
 	int i;
